Fix warnings

+++ dftcalc/dftcalc.cpp
@@ -975,7 +975,7 @@ int main(int argc, char** argv) {
 			printHelp = true;
 			break;
 		} else if (!strncmp(argv[argi], "-v", 2)) {
-			for (int i = 1; i < strlen(argv[argi]); i++) {
+			for (size_t i = 1; i < strlen(argv[argi]); i++) {
 				if (argv[argi][i] == 'v') {
 					++verbosity;
 				} else {
